+++ mlbusinesscomponents/src/main/java/com/mercadolibre/android/mlbusinesscomponents/components/touchpoint/view/flex_cover_carousel/FlexCoverCarouselComponent.kt
@@ -1,7 +1,10 @@
 package com.mercadolibre.android.mlbusinesscomponents.components.touchpoint.view.flex_cover_carousel
 
 import android.content.Context
+import android.graphics.Rect
 import android.util.AttributeSet
+import android.view.View
+import android.view.ViewGroup
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.PagerSnapHelper
 import androidx.recyclerview.widget.RecyclerView
@@ -10,7 +13,11 @@ import com.mercadolibre.android.mlbusinesscomponents.components.touchpoint.callb
 import com.mercadolibre.android.mlbusinesscomponents.components.touchpoint.domain.model.AdditionalEdgeInsets
 import com.mercadolibre.android.mlbusinesscomponents.components.touchpoint.domain.model.flex_cover_carousel.FlexCoverCard
 import com.mercadolibre.android.mlbusinesscomponents.components.touchpoint.domain.model.flex_cover_carousel.FlexCoverCarouselResponse
+import com.mercadolibre.android.mlbusinesscomponents.components.touchpoint.tracking.MLBusinessTouchpointTracker
+import com.mercadolibre.android.mlbusinesscomponents.components.touchpoint.tracking.TouchpointTrackeable
+import com.mercadolibre.android.mlbusinesscomponents.components.touchpoint.tracking.print.TouchpointPrintProvider
 import com.mercadolibre.android.mlbusinesscomponents.components.utils.ScaleUtils
+import com.mercadolibre.android.mlbusinesscomponents.components.utils.TrackingUtils
 
 class FlexCoverCarouselComponent @JvmOverloads constructor(
     context: Context,
@@ -28,12 +35,20 @@ class FlexCoverCarouselComponent @JvmOverloads constructor(
     private var wasAtLastElement: Boolean? = null
     private var normalPadding = 0
     private var biggerPadding = 0
+    private lateinit var tracker: MLBusinessTouchpointTracker
+    private lateinit var printProvider: TouchpointPrintProvider
+    private val recyclerRect by lazy { Rect() }
 
-    override fun bind(model: FlexCoverCarouselResponse, onClickCallback: OnClickCallback?) {
+    override fun bind(
+        model: FlexCoverCarouselResponse,
+        onClickCallback: OnClickCallback?,
+        tracker: MLBusinessTouchpointTracker?,
+        printProvider: TouchpointPrintProvider
+    ) {
         if (!isInitialized) {
             initialize()
         }
-        presenter.bind(model, onClickCallback)
+        presenter.bind(model, onClickCallback, tracker, printProvider)
     }
 
     override fun notifyPadding(additionalEdgeInsets: AdditionalEdgeInsets) {
@@ -41,11 +56,19 @@ class FlexCoverCarouselComponent @JvmOverloads constructor(
         biggerPadding = ScaleUtils.getPxFromDp(context, additionalEdgeInsets.right.toFloat()).toInt()
     }
 
+    override fun setTracker(tracker: MLBusinessTouchpointTracker?) {
+        tracker?.let { this.tracker = it }
+    }
+
+    override fun setPrintProvider(printProvider: TouchpointPrintProvider) {
+        this.printProvider = printProvider
+    }
+
     override fun setCards(cards: List<FlexCoverCard>, onClickCallback: OnClickCallback?) {
         cardsAdapter?.let {
             it.updateData(cards)
         } ?: let {
-            cardsAdapter = FlexCoverCarouselComponentAdapter(cards, onClickCallback)
+            cardsAdapter = FlexCoverCarouselComponentAdapter(cards, onClickCallback, tracker)
             adapter = cardsAdapter
             val decorator =
                 FlexCoverCarouselComponentDecorator(resources.getDimensionPixelSize(R.dimen.ui_1m))
@@ -64,7 +87,11 @@ class FlexCoverCarouselComponent @JvmOverloads constructor(
 
     private fun createScrollListener() = object : OnScrollListener() {
 
-        override fun onScrollStateChanged(recyclerView: RecyclerView, newState: Int) {}
+        override fun onScrollStateChanged(recyclerView: RecyclerView, newState: Int) {
+            if (newState == SCROLL_STATE_IDLE) {
+                print(recyclerView)
+            }
+        }
 
         override fun onScrolled(recyclerView: RecyclerView, dx: Int, dy: Int) {
             super.onScrolled(recyclerView, dx, dy)
@@ -96,4 +123,26 @@ class FlexCoverCarouselComponent @JvmOverloads constructor(
     }
 
     private fun getLastItemIndex() = linearLayoutManager.itemCount - 1
+
+    fun print(recyclerView: RecyclerView) {
+        recyclerView.getHitRect(recyclerRect)
+        findData(recyclerView)
+        TrackingUtils.trackPrint(tracker, printProvider)
+    }
+
+    private fun findData(viewGroup: ViewGroup) {
+        for (i in 0 until viewGroup.childCount) {
+            val child = viewGroup.getChildAt(i)
+            if (child is ViewGroup && child !is TouchpointTrackeable) {
+                findData(child)
+            }
+            if (shouldTrackPrint(child)) {
+                printProvider.accumulateData((child as TouchpointTrackeable).tracking)
+            }
+        }
+    }
+
+    private fun shouldTrackPrint(child: View): Boolean {
+        return child is TouchpointTrackeable && child.getLocalVisibleRect(recyclerRect)
+    }
 }