+++ mlbusinesscomponents/src/main/java/com/mercadolibre/android/mlbusinesscomponents/components/touchpoint/view/flex_cover_carousel/FlexCoverCarouselComponentPresenter.kt
@@ -2,13 +2,22 @@ package com.mercadolibre.android.mlbusinesscomponents.components.touchpoint.view
 
 import com.mercadolibre.android.mlbusinesscomponents.components.touchpoint.callback.OnClickCallback
 import com.mercadolibre.android.mlbusinesscomponents.components.touchpoint.domain.model.flex_cover_carousel.FlexCoverCarouselResponse
+import com.mercadolibre.android.mlbusinesscomponents.components.touchpoint.tracking.MLBusinessTouchpointTracker
+import com.mercadolibre.android.mlbusinesscomponents.components.touchpoint.tracking.print.TouchpointPrintProvider
 
 class FlexCoverCarouselComponentPresenter(val view: FlexCoverCarouselComponentInterface) {
 
     private var model: FlexCoverCarouselResponse? = null
 
-    fun bind(model: FlexCoverCarouselResponse, onClickCallback: OnClickCallback?) {
+    fun bind(
+        model: FlexCoverCarouselResponse,
+        onClickCallback: OnClickCallback?,
+        tracker: MLBusinessTouchpointTracker?,
+        printProvider: TouchpointPrintProvider
+    ) {
         this.model = model
+        view.setTracker(tracker)
         view.setCards(model.items, onClickCallback)
+        view.setPrintProvider(printProvider)
     }
 }